pyproject.toml
[project]
name = "k8s-pilot"
version = "0.1.0"
description = "k8s-pilot is a tool to help you manage Kubernetes clusters."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "kubernetes>=32.0.1",
    "mcp[cli]>=1.6.0",
    "ruff>=0.11.5",
    "pyyaml",
]

[tool.mcp]
entrypoint = "main.py"